+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.6.0 (IN PROGRESS)
 
 * React deps should be as lenient as possible.
+* Add request policy CRU v1. Fixes UIREQ-181.
 
 ## 1.5.0 (https://github.com/folio-org/ui-circulation/tree/v1.5.0) (2019-01-25)
 [Full Changelog](https://github.com/folio-org/ui-circulation/compare/v1.4.0...v1.5.0)